trying to figure out why 0-print_dlistint.c not compling in checker sandbox

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -10,7 +10,7 @@ size_t print_dlistint(const dlistint_t *h)
 {
 	const dlistint_t *tmp;
 
-	int lenght = 0;
+	size_t lenght = 0;
 
 	if (h == NULL)
 		return (0);
